requirements.txt
pytest==8.3.4
selenium==4.27.1
retry == 0.9.2
allure-pytest == 2.13.5
undetected-chromedriver == 3.5.5
